refactor: move colorSchemeOptions file
