+++ src/widgetLibrary/ButtonWidget/panelConfig.tsx
@@ -5,7 +5,7 @@ import {
   HorizontalFullIcon,
   HorizontalStartIcon,
 } from "@illa-design/react"
-import { colorSchemeOptions } from "@/widgetLibrary/colorSchemeOptions"
+import { colorSchemeOptions } from "@/widgetLibrary/PublicSector/colorSchemeOptions"
 import i18n from "@/i18n/config"
 
 export const BUTTON_PANEL_CONFIG: PanelConfig[] = [
