+++ src/widgetLibrary/TextWidget/panelConfig.tsx
@@ -7,7 +7,7 @@ import {
   VerticalCenterIcon,
   VerticalEndIcon,
 } from "@illa-design/react"
-import { colorSchemeOptions } from "@/widgetLibrary/colorSchemeOptions"
+import { colorSchemeOptions } from "@/widgetLibrary/PublicSector/colorSchemeOptions"
 import i18n from "@/i18n/config"
 
 export const TEXT_PANEL_CONFIG: PanelConfig[] = [
